pyproject.toml
[project]
name = "market-pulse"
version = "0.1.0"
description = "LLM based job market toolkit to optimize user profile"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "ibm-cloudant>=0.9.1",
    "ibm-cloud-sdk-core>=3.20.0",
    "httpx>=0.26.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
